Restructure for update-cart logic
